% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{assert_correctly_formed_pattern_expression}
\alias{assert_correctly_formed_pattern_expression}
\title{Raise an error if a match expression is malformed.}
\usage{
assert_correctly_formed_pattern_expression(match_expr)
}
\arguments{
\item{match_expr}{The match expression}
}
\description{
Raise an error if a match expression is malformed.
}
