% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\docType{data}
\name{bind}
\alias{bind}
\title{Bind variables to pattern-matched expressions.}
\format{}
\usage{
bind[variables] <- values
}
\description{
The \code{bind} object itself doesn't do anything. It simply exists in order to define
notation for binding variabels using the sub-script operator.
}
\examples{
bind[x, y] <- c(2,4)
x == 2
y == 4

llist := NIL | CONS(car, cdr : llist)
L <- CONS(1, CONS(2, CONS(3, NIL)))
bind[CONS(first, CONS(second, rest))] <- L
first == 1
second == 2

}
\keyword{datasets}
