% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{construction_printer}
\alias{construction_printer}
\title{Print a constructed value}
\usage{
construction_printer(x, ...)
}
\arguments{
\item{x}{Object to print}

\item{...}{Additional parameters; not used.}
}
\description{
Print a constructed value
}
