% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{deparse_construction}
\alias{deparse_construction}
\title{Create a string representation from a constructed object}
\usage{
deparse_construction(x, ...)
}
\arguments{
\item{x}{The object to translate into a string}

\item{...}{Additional parameters; not used.}
}
\value{
A string representation of \code{object}
}
\description{
Create a string representation from a constructed object
}
