% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{make_args_list}
\alias{make_args_list}
\title{Construct a pair-list of arguments that can be used to create a new function}
\usage{
make_args_list(args)
}
\arguments{
\item{args}{List of variable names.}
}
\value{
A pair list that can be used with rlang::new_function.
}
\description{
Given a list of argument names, construct a list of arguments with empty defaults.
}
