% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{process_alternatives}
\alias{process_alternatives}
\title{Goes through a list of |-separated expressions and define them as constructors}
\usage{
process_alternatives(constructors, data_type_name, env)
}
\arguments{
\item{constructors}{The constructs specification}

\item{data_type_name}{The type the constructor should generate}

\item{env}{The environment where we define the constructor}
}
\description{
Goes through a list of |-separated expressions and define them as constructors
}
