% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{process_arg}
\alias{process_arg}
\title{Build a tibble form a list of constructor arguments.}
\usage{
process_arg(argument)
}
\arguments{
\item{argument}{The argument provided to a constructor in its definition}
}
\value{
A tibble with a single row, the first column holds the argument name,
the second its type.
}
\description{
Build a tibble form a list of constructor arguments.
}
