% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{process_arguments}
\alias{process_arguments}
\title{Construct a tibble from all the arguments of a constructor}
\usage{
process_arguments(constructor_arguments)
}
\arguments{
\item{constructor_arguments}{The arguments provided in the constructor
specification}
}
\value{
The arguments represented as a tibble. The first column contain
argument names, the second their types.
}
\description{
Construct a tibble from all the arguments of a constructor
}
