% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{process_constructor_function}
\alias{process_constructor_function}
\title{Create a function constructor and put it in an environment.}
\usage{
process_constructor_function(constructor, data_type_name, env)
}
\arguments{
\item{constructor}{The construct specification}

\item{data_type_name}{The type the constructor should generate}

\item{env}{The environment where we define the constructor}
}
\description{
Create a function constructor and put it in an environment.
}
