% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{[<-.pmatch_bind}
\alias{[<-.pmatch_bind}
\title{Bind variables to pattern-matched expressions.}
\usage{
\method{[}{pmatch_bind}(dummy, ...) <- value
}
\arguments{
\item{dummy}{The \code{bind} object. Only used to dispatch to the right
subscript operator.}

\item{...}{Patterns to assign to.}

\item{value}{Actual values to assign}
}
\description{
The \code{bind} object itself doesn't do anything. It simply exists in order
to define notation for binding variables using the sub-script operator.
}
\examples{
bind[x, y] <- c(2,4)
x == 2
y == 4

llist := NIL | CONS(car, cdr : llist)
L <- CONS(1, CONS(2, CONS(3, NIL)))
bind[CONS(first, CONS(second, rest))] <- L
first == 1
second == 2

}
