% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{test_pattern_rec}
\alias{test_pattern_rec}
\title{Recursive comparison of expression and pattern.}
\usage{
test_pattern_rec(escape, expr, test_expr, eval_env, match_env)
}
\arguments{
\item{escape}{Continuation from callCC, used to escape if we cannot
match.}

\item{expr}{The expression to match again.}

\item{test_expr}{The pattern we are trying to match.}

\item{eval_env}{The environment where we get constructors from.}

\item{match_env}{The environment to put matched variables in.}
}
\value{
An environment containing bound variables from the expression, if
matching. If the pattern doesn't match, the function escapes through the
\code{escape} continuation.
}
\description{
Recursive comparison of expression and pattern.
}
