% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{`..`}
\alias{`..`}
\alias{..}
\title{Function for specifying a sequence of patterns/expressions}
\usage{
..(...)
}
\arguments{
\item{...}{The patterns/expressions to combine}
}
\description{
This function is used when you want to test for more than one pattern
in parallel
}
