% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailr-transforms.R
\name{transform_call.tailr_pmatch_cases}
\alias{transform_call.tailr_pmatch_cases}
\title{Transformation hook for tailr}
\usage{
transform_call.tailr_pmatch_cases(fun, expr)
}
\arguments{
\item{fun}{The function used to dispatch the generic function on. This would be the \code{\link{cases}} function.}

\item{expr}{The call expression to transform.}
}
\value{
Transformed expression.
}
\description{
Transformation hook for tailr
}
