% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_cases_call}
\alias{transform_cases_call}
\title{Recursive function for transforming a call \code{cases}.}
\usage{
transform_cases_call(expr, ...)
}
\arguments{
\item{expr}{The expression to transform.}

\item{...}{Additional callback arguments to make this work with \code{foolbox}}
}
\value{
Updated expression.
}
\description{
Recursive function for transforming a call \code{cases}.
}
