% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_cases_function}
\alias{transform_cases_function}
\title{Transform a function containing a \code{cases} call into one that
instead has if-statements.}
\usage{
transform_cases_function(fun)
}
\arguments{
\item{fun}{A function}
}
\value{
Another function with a transformed body
}
\description{
Transform a function containing a \code{cases} call into one that
instead has if-statements.
}
\seealso{
cases
}
