% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_cases_function_rec}
\alias{transform_cases_function_rec}
\title{Recursive function for transforming a function that uses \code{cases}.}
\usage{
transform_cases_function_rec(expr)
}
\arguments{
\item{expr}{The expression to transform.}
}
\value{
Updated expression.
}
\description{
Recursive function for transforming a function that uses \code{cases}.
}
