% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\docType{data}
\name{bind}
\alias{bind}
\title{Dummy object used for generic function dispatching.}
\format{An object of class \code{pmatch_bind} of length 1.}
\usage{
bind
}
\description{
Dummy object used for generic function dispatching.
}
\keyword{datasets}
